graphic — clear version — копия/requirements.txt
# Python dependencies for the Scientific Visualization Project
# Compatible with Python 3.8+
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Symbolic mathematics and LaTeX parsing
sympy>=1.9

# Plotting and visualization
matplotlib>=3.4.0

# Data processing and Excel support
pandas>=1.3.0
openpyxl>=3.0.0

# Configuration files
PyYAML>=5.4.0

# Optional: For better performance
# numba>=0.54.0  # Uncomment if you need JIT compilation
